+++ kyuubi-common/src/main/scala/org/apache/kyuubi/service/authentication/UserDefinedEngineSecureAccessSecretProvider.scala
@@ -17,32 +17,14 @@
 
 package org.apache.kyuubi.service.authentication
 
-import java.nio.charset.StandardCharsets
-
 import org.apache.kyuubi.config.KyuubiConf
-import org.apache.kyuubi.ha.HighAvailabilityConf.HA_ZK_ENGINE_SECRET_NODE
-import org.apache.kyuubi.ha.client.ZooKeeperClientProvider
 
-class ZooKeeperEngineSecureAccessProviderImpl extends EngineSecureAccessProvider {
-  import ZooKeeperClientProvider._
+class UserDefinedEngineSecureAccessSecretProvider extends EngineSecureAccessSecretProvider {
+  override def initialize(kyuubiConf: KyuubiConf): Unit = {}
 
-  private var conf: KyuubiConf = _
-
-  override def initialize(conf: KyuubiConf): Unit = {
-    this.conf = conf
-  }
-
-  override def supportSecureAccess: Boolean = {
-    conf.get(HA_ZK_ENGINE_SECRET_NODE).nonEmpty
-  }
+  override def supportSecureAccess: Boolean = true
 
   override def getSecret(): String = {
-    var secret: String = ""
-    conf.get(HA_ZK_ENGINE_SECRET_NODE).map { zkNode =>
-      withZkClient(conf) { zkClient =>
-        secret = new String(zkClient.getData.forPath(zkNode), StandardCharsets.UTF_8)
-      }
-    }
-    secret
+    "ENGINE____SECRET"
   }
 }
